requirements.txt
celery>=3.1.5,<3.2
django>=1.3,<1.8
djangorestframework>=2.3,<2.4
factory-boy==2.8.1
mock==1.0.1
six
jsonpickle>=0.8.0
